requirements.txt
fastapi
uvicorn[standard]
requests
sentence-transformers
torch
psycopg2-binary
pgvector
